+++ src/lib/blit.c
@@ -175,6 +175,7 @@ rgb_diff(unsigned r1, unsigned g1, unsigned b1, unsigned r2, unsigned g2, unsign
   distance += r1 > r2 ? r1 - r2 : r2 - r1;
   distance += g1 > g2 ? g1 - g2 : g2 - g1;
   distance += b1 > b2 ? b1 - b2 : b2 - b1;
+//fprintf(stderr, "RGBDIFF %u %u %u %u %u %u: %u\n", r1, g1, b1, r2, g2, b2, distance);
   return distance;
 }
 
@@ -258,28 +259,51 @@ quadrant_solver(uint32_t tl, uint32_t tr, uint32_t bl, uint32_t br,
 //fprintf(stderr, "mindiff: %u[%zu] fore: %08x back: %08x %d+%d/%d+%d\n", mindiff, mindiffidx, *fore, *back, qd->pair[0], qd->pair[1], qd->others[0], qd->others[1]);
   const char* egc = qd->egc;
   // break down the excluded pair and lerp
-  unsigned r0, r1, g0, g1, b0, b1;
+  unsigned r0, r1, r2, g0, g1, g2, b0, b1, b2;
   unsigned roth, goth, both, rlerp, glerp, blerp;
+  channel_rgb8(*back, &roth, &goth, &both);
+  channel_rgb8(*fore, &rlerp, &glerp, &blerp);
+//fprintf(stderr, "rgbs: %02x %02x %02x / %02x %02x %02x\n", r0, g0, b0, r1, g1, b1);
+  // get diffs of the excluded two from both lerps
   channel_rgb8(colors[qd->others[0]], &r0, &g0, &b0);
   channel_rgb8(colors[qd->others[1]], &r1, &g1, &b1);
-  channel_rgb8(*fore, &rlerp, &glerp, &blerp);
-  channel_rgb8(*back, &roth, &goth, &both);
-//fprintf(stderr, "rgbs: %02x %02x %02x / %02x %02x %02x\n", r0, g0, b0, r1, g1, b1);
-  diffs[0] = rgb_diff(r0, g0, b0, rlerp, glerp, blerp);
-  diffs[1] = rgb_diff(r0, g0, b0, roth, goth, both);
-  diffs[2] = rgb_diff(r1, g1, b1, rlerp, glerp, blerp);
-  diffs[3] = rgb_diff(r1, g1, b1, roth, goth, both);
-//fprintf(stderr, "diffs: %08x %08x %08x %08x\n", diffs[0], diffs[1], diffs[2], diffs[3]);
-  if(diffs[0] < diffs[1] && diffs[0] < diffs[2]){
-    egc = qd->oth0egc;
-    *back = colors[qd->others[1]];
-    *fore = trilerp(colors[qd->pair[0]], colors[qd->pair[1]], colors[qd->others[0]]);
-//fprintf(stderr, "swap 1 %08x %08x\n", *fore, *back);
-  }else if(diffs[2] < diffs[3]){
-    egc = qd->oth1egc;
-    *back = colors[qd->others[0]];
-    *fore = trilerp(colors[qd->pair[0]], colors[qd->pair[1]], colors[qd->others[1]]);
-//fprintf(stderr, "swap 2 %08x %08x\n", *fore, *back);
+  diffs[0] = rgb_diff(r0, g0, b0, roth, goth, both);
+  diffs[1] = rgb_diff(r1, g1, b1, roth, goth, both);
+  diffs[2] = rgb_diff(r0, g0, b0, rlerp, glerp, blerp);
+  diffs[3] = rgb_diff(r1, g1, b1, rlerp, glerp, blerp);
+  // get diffs of the included two from their lerp
+  channel_rgb8(colors[qd->pair[0]], &r0, &g0, &b0);
+  channel_rgb8(colors[qd->pair[1]], &r1, &g1, &b1);
+  diffs[4] = rgb_diff(r0, g0, b0, rlerp, glerp, blerp);
+  diffs[5] = rgb_diff(r1, g1, b1, rlerp, glerp, blerp);
+  unsigned curdiff = diffs[0] + diffs[1] + diffs[4] + diffs[5];
+  // it might be better to combine three, and leave one totally unchanged.
+  // propose a trilerps; we only need consider the member of the excluded pair
+  // closer to the primary lerp. recalculate total diff; merge if lower.
+  if(diffs[2] < diffs[3]){
+    unsigned tri = trilerp(colors[qd->pair[0]], colors[qd->pair[1]], colors[qd->others[0]]);
+    channel_rgb8(colors[qd->others[0]], &r2, &g2, &b2);
+    channel_rgb8(tri, &roth, &goth, &both);
+    if(rgb_diff(r0, g0, b0, roth, goth, both) +
+       rgb_diff(r1, g1, b1, roth, goth, both) +
+       rgb_diff(r2, g2, b2, roth, goth, both) < curdiff){
+      egc = qd->oth0egc;
+      *back = colors[qd->others[1]];
+      *fore = tri;
+    }
+//fprintf(stderr, "quadblitter swap type 1\n");
+  }else{
+    unsigned tri = trilerp(colors[qd->pair[0]], colors[qd->pair[1]], colors[qd->others[1]]);
+    channel_rgb8(colors[qd->others[1]], &r2, &g2, &b2);
+    channel_rgb8(tri, &roth, &goth, &both);
+    if(rgb_diff(r0, g0, b0, roth, goth, both) +
+       rgb_diff(r1, g1, b1, roth, goth, both) +
+       rgb_diff(r2, g2, b2, roth, goth, both) < curdiff){
+      egc = qd->oth1egc;
+      *back = colors[qd->others[0]];
+      *fore = tri;
+    }
+//fprintf(stderr, "quadblitter swap type 2\n");
   }
   return egc;
 }
@@ -489,20 +513,20 @@ generalerp(unsigned rsum, unsigned gsum, unsigned bsum, int count){
 // of pixels that minimizes total source distance from the resulting lerps.
 static const char*
 sex_solver(const uint32_t rgbas[6], uint64_t* channels, bool blendcolors){
-	static const char* sex[69] = {
-	  " ", "🬀", "🬁", "🬃", "🬇", "🬏", "🬞", "🬂",
-	  "🬄", "🬈", "🬐", "🬟", "🬅", "🬉", "🬑", "🬠",
-	  "🬋", "🬓", "🬢", "🬖", "🬦", "🬭", "🬆", "🬊",
-	  "🬒", "🬡", "🬌", "▌", "🬣", "🬗", "🬧", "🬍",
-	  "█", "🬻", "🬺", "🬸", "🬴", "🬬", "🬝", "🬹",
-	  "🬷", "🬳", "🬫", "🬜", "🬶", "🬲", "🬪", "🬛",
-	  "🬰", "🬨", "🬙", "🬥", "🬕", "🬎", "🬵", "🬱",
-	  "🬩", "🬚", "🬯", "▐", "🬘", "🬤", "🬔", "🬮",
-    "n", "i", "c", "e"
-	};
   // each element within the set of 64 has an inverse element within the set,
-  // for which we will calculate the same total differences, so just handle the
-  // first 32, and then assign fg to whichever cluster is larger.
+  // for which we would calculate the same total differences, so just handle
+  // the first 32. the partition[] bit masks represent combinations of
+  // sextants, and their indices correspond to sex[].
+  static const char* sex[32] = {
+    " ", "🬀", "🬁", "🬃", "🬇", "🬏", "🬞", "🬂", // 0..7
+
+    "🬄", "🬈", "🬐", "🬟", "🬅", "🬉", "🬑", "🬠", // 8..15
+
+    "🬋", "🬓", "🬢", "🬖", "🬦", "🬭", "🬆", "🬊", // 16..23
+
+    "🬒", "🬡", "🬌", "▌", "🬣", "🬗", "🬧", "🬍", // 24..31
+
+  };
   static const unsigned partitions[32] = {
     0, // 1 way to arrange 0
     1, 2, 4, 8, 16, 32, // 6 ways to arrange 1
@@ -533,9 +557,9 @@ sex_solver(const uint32_t rgbas[6], uint64_t* channels, bool blendcolors){
         bsum1 += ncpixel_b(rgbas[mask]);
       }
     }
-//fprintf(stderr, "sum0: %u/%u/%u sum1: %u/%u/%u insum: %d\n", rsum0, gsum0, bsum0, rsum1, gsum1, bsum1, insum);
     uint32_t l0 = generalerp(rsum0, gsum0, bsum0, insum);
     uint32_t l1 = generalerp(rsum1, gsum1, bsum1, 6 - insum);
+//fprintf(stderr, "sum0: %06x sum1: %06x insum: %d\n", l0 & 0xffffffu, l1 & 0xffffffu, insum);
     uint32_t totaldiff = 0;
     for(unsigned mask = 0 ; mask < 6 ; ++mask){
       unsigned r, g, b;
@@ -544,11 +568,12 @@ sex_solver(const uint32_t rgbas[6], uint64_t* channels, bool blendcolors){
       }else{
         channel_rgb8(l1, &r, &g, &b);
       }
-      totaldiff += rgb_diff(ncpixel_r(rgbas[mask]), ncpixel_g(rgbas[mask]),
-                            ncpixel_b(rgbas[mask]), r, g, b);
+      uint32_t rdiff = rgb_diff(ncpixel_r(rgbas[mask]), ncpixel_g(rgbas[mask]),
+                                ncpixel_b(rgbas[mask]), r, g, b);
+      totaldiff += rdiff;
 //fprintf(stderr, "mask: %u totaldiff: %u insum: %d (%08x / %08x)\n", mask, totaldiff, insum, l0, l1);
     }
-//fprintf(stderr, "bits: %u %zu totaldiff: %u best: %u (%d)\n", partitions[glyph], glyph, totaldiff, mindiff, best);
+//fprintf(stderr, "bits: %u %zu totaldiff: %f best: %f (%d)\n", partitions[glyph], glyph, totaldiff, mindiff, best);
     if(totaldiff < mindiff){
       mindiff = totaldiff;
       best = glyph;
@@ -560,7 +585,7 @@ sex_solver(const uint32_t rgbas[6], uint64_t* channels, bool blendcolors){
     }
   }
 //fprintf(stderr, "solved for best: %d (%u)\n", best, mindiff);
-  assert(best >= 0 && best < 64);
+  assert(best >= 0 && best < 32);
   if(blendcolors){
     channels_set_fg_alpha(channels, CELL_ALPHA_BLEND);
     channels_set_bg_alpha(channels, CELL_ALPHA_BLEND);